refactor: choice wizard state controller 분리
